notifier_function/requirements.txt
functions-framework==3.*
google-cloud-pubsub>=0.28.1
twilio==7.15.*